% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_train_predict.R
\name{avg_prob}
\alias{avg_prob}
\title{Computes average confidence scores}
\usage{
avg_prob(data)
}
\arguments{
\item{data}{tibble with at least a common_seed column and a prediction_output
column The prediction_output column is a list. Each element contains a
tibble with predictions and covariates.}
}
\value{
an object (tibble) with all the covariates that identify the
vessel-years, and an average predicted probability for each vessel-year
and common seed
}
\description{
For each common seed and vessel-year, it computes the average
confidence score.
}
