% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_downsample.R
\name{bag_downsample}
\alias{bag_downsample}
\title{Apply downsample step to data recipe across defined seeds/bags}
\usage{
bag_downsample(bag_runs, fl_rec, down_sample_ratio)
}
\arguments{
\item{bag_runs}{Tibble defining bag numbers and seeds}

\item{fl_rec}{Model recipe}

\item{down_sample_ratio}{See \code{under_ratio} \code{\link[themis:step_downsample]{themis::step_downsample()}}.
Downsampling ratio to balance the number of positive and unlabelled cases.
Defaults to 1 (1:1 ratio)}
}
\value{
Tibble with data recipe and downsample ratio.
}
\description{
Apply downsample step to data recipe across defined seeds/bags
}
