% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrated_threshold.R
\name{calibrated_threshold}
\alias{calibrated_threshold}
\title{Computes threshold for offender classification}
\usage{
calibrated_threshold(
  data,
  steps = 1000,
  plotting = FALSE,
  filename = NULL,
  threshold = seq(0, 0.99, by = 0.01),
  eps = 0.01
)
}
\arguments{
\item{data}{data frame. Needs to have a .pred_1 column with predictions
and a known_offender column with 0 for unlabeled and 1 for positive
(offender)}

\item{steps}{number of locations at which to compute D}

\item{plotting}{if TRUE, a D vs. alpha plot is generated}

\item{filename}{if plotting is TRUE, a filename with path is required}

\item{threshold}{potential thresholds to test}

\item{eps}{accepted difference (tolerance) between alpha and the actual
proportion of positives for a given threshold}
}
\value{
a threshold to use
}
\description{
Computes threshold for offender classification based on alpha
(from the dedpul_estimation function—see reference): which threshold would
achieve an alpha or proportion of positives within the unlabeled (more or
less) equal to alpha?
}
\details{
For more details on the algorithm, please see the reference.
}
\references{
D. Ivanov, "DEDPUL: Difference-of-Estimated-Densities-based
Positive-Unlabeled Learning," 2020 19th IEEE International Conference on
Machine Learning and Applications (ICMLA), 2020, pp. 782-790, doi:
10.1109/ICMLA51294.2020.00128.
}
