% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_estimate.R
\name{conf_estimate}
\alias{conf_estimate}
\title{Compute a confidence level estimate}
\usage{
conf_estimate(predicted_df, data, threshold)
}
\arguments{
\item{predicted_df}{dataframe of one row containing the predicted class and
the ID of the vessel, that should match the IDs in data}

\item{data}{dataframe with all the random forest scores}

\item{threshold}{threshold to classify the scores into 0 or 1}
}
\value{
confidence level estimates
}
\description{
Computing confidence level estimates for the predicted class
based on a beta distribution fitted on the random forest scores and
computing the probability mass below or above the classification threshold,
if the class is 0 or 1, respectively
}
