% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_bag_downsample.R
\name{dev_bag_downsample}
\alias{dev_bag_downsample}
\title{Apply downsample step to data recipe across defined seeds/bags}
\usage{
dev_bag_downsample(bag_runs, fl_rec, down_sample_ratio)
}
\arguments{
\item{bag_runs}{Tibble defining bag numbers and seeds}

\item{fl_rec}{Recipes data recipe}

\item{down_sample_ratio}{See under_ratio in ?themis::step_downsample}
}
\value{
Tibble with data recipe and downsample ratio.
}
\description{
Apply downsample step to data recipe across defined seeds/bags
}
