% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ml_load.R
\name{dev_ml_load}
\alias{dev_ml_load}
\title{Loading trained rf models from disk based on the cross validation set up defined}
\usage{
dev_ml_load(cv_setup, save_dir)
}
\arguments{
\item{cv_setup}{List containing cv_folds (analysis/assessment), model workflow and seed/bag identifiers. Output from ?cv_setup}

\item{save_dir}{Directory to save trained models otherwise skip saving when NULL}
}
\value{
List containing:
Trained random forest models
}
\description{
Loading trained rf models from disk based on the cross validation set up defined
}
