% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ml_predict.R
\name{dev_ml_predict}
\alias{dev_ml_predict}
\title{Predicting over new data using previously trained RF models}
\usage{
dev_ml_predict(trained_models, new_data, free_cores, parallel_plan)
}
\arguments{
\item{trained_models}{List of RF models (as returned by dev_ml_train; dev_ml_load)}

\item{new_data}{New data over which to apply trained models to generate predictions}

\item{free_cores}{Number of available cores. Add more if you need to do many things at the same time}

\item{parallel_plan}{Parallelization strategy Options: multisession (if running RStudio), multicore (Linux, Mac and plain R) or psock (if multisession is not working well and you need to try something else)}
}
\value{
Dataframe with predicted values for each RF model
}
\description{
Predicting over new data using previously trained RF models
}
