% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ml_tune.R
\name{dev_ml_tune}
\alias{dev_ml_tune}
\title{Tune random forest hyper-parameters specified during dev_rf_setup}
\usage{
dev_ml_tune(
  training_data,
  fl_rec,
  rf_spec,
  num_folds,
  num_bags,
  num_seeds,
  down_sample_ratio,
  tune_parameters = NULL,
  grid = NULL,
  parallel_plan,
  free_cores
)
}
\arguments{
\item{training_data}{Dataset over which tune model hyperparameters}

\item{fl_rec}{Model recipe}

\item{rf_spec}{Random forest classifier specifications}

\item{num_folds}{Number of cross validation folds}

\item{num_bags}{Number of bags}

\item{num_seeds}{Number of common seeds}

\item{down_sample_ratio}{See under_ratio in ?themis::step_downsample. To reduce the weight of the unlabeled cases in the model, we randomly
downsampled them in the training set with a 1-1 ratio, i.e. the number of
positive and unlabeled cases used for training would be equal}

\item{tune_parameters}{String defining parameter (or parameters) over which to perform tuning. Default to NULL which perform tuning across all hyperparameters in rf_spec}

\item{grid}{number of grid random values for combinations of hyperparameters per bag, or grid of values to test}

\item{parallel_plan}{type of parallelization to run (multicore, multisession
or psock - this last one may need calling libraries inside)}

\item{free_cores}{number of free cores to leave out of parallelization}
}
\value{
Data frame of train cross-validated datasets
}
\description{
Tune random forest hyper-parameters specified during dev_rf_setup
}
