% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_cvsetup.R
\name{fl_cvsetup}
\alias{fl_cvsetup}
\title{Setting up data structure}
\usage{
fl_cvsetup(
  training_data,
  fl_rec,
  rf_spec,
  num_seeds,
  num_bags,
  num_folds,
  down_sample_ratio,
  group_var = "source_id"
)
}
\arguments{
\item{training_data}{Dataset over which to generate CV folds and tuning grid}

\item{fl_rec}{Model recipe}

\item{rf_spec}{Random forest classifier specifications}

\item{num_seeds}{Number of common seeds}

\item{num_bags}{Number of bags}

\item{num_folds}{Number of cross validation folds}

\item{down_sample_ratio}{See \code{under_ratio} \code{\link[themis:step_downsample]{themis::step_downsample()}}.
Downsampling ratio to balance the number of positive and unlabelled cases.
Defaults to 1 (1:1 ratio)}

\item{group_var}{A variable in data (single character or name) used for
grouping observations with the same value to assign cases to train or test
sets within a fold using \code{\link[rsample:group_vfold_cv]{rsample::group_vfold_cv()}}.}
}
\value{
List object containing cv_folds (analysis/assessment), model
workflow and seed/bag identifiers
}
\description{
Sets up the data structure to train the RF model based on a number of folds,
bags and common seeds.
}
