% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_recall.R
\name{fl_recall}
\alias{fl_recall}
\title{Computes recall for assessment sets}
\usage{
fl_recall(data)
}
\arguments{
\item{data}{tibble with at least a prediction output column (\code{pred_class})
and a \code{known_offender} column (whether the vessel was identified as
an offender by reports).}
}
\value{
recall value
}
\description{
Computes recall, for assessment sets (in model versions that
did not use them for training)
}
