% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_train.R
\name{fl_train}
\alias{fl_train}
\title{Training forced labor random forest model.}
\usage{
fl_train(
  cv_setup = cv_df[[1]],
  rf_spec = rf_setup$rf_spec,
  holdout = NULL,
  save_dir = NULL
)
}
\arguments{
\item{cv_setup}{List containing cv_folds (analysis/assessment), model workflow and seed/bag identifiers. Output from ?cv_setup}

\item{rf_spec}{Random forest classifier specifications}

\item{holdout}{Optional. Test dataset, not used for model training.}

\item{save_dir}{Directory to save trained models otherwise skip saving when NULL}
}
\value{
List containing:
Trained random forest models
Tibble with predicted probabilities
}
\description{
Training forced labor random forest model.
}
