% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcedlabor-package.R
\docType{package}
\name{forcedlabor-package}
\alias{forcedlabor}
\alias{forcedlabor-package}
\title{forcedlabor: Global Fishing Watch's forced labor risk machine learning algorithm}
\description{
forcedlabor is an R package containing the codes to classify the risk of forced labor on fishing vessels via a machine learning algorithm trained on a combination of known forced labor cases and AIS data predictors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GlobalFishingWatch/forcedlabor}
  \item Report bugs at \url{https://github.com/GlobalFishingWatch/forcedlabor/issues}
}

}
\author{
\strong{Maintainer}: Rocío Joo \email{rocio.joo@globalfishingwatch.org} (\href{https://orcid.org/0000-0003-0319-4210}{ORCID})

Authors:
\itemize{
  \item Guillermo Martín González \email{guillermo.martin@globalfishingwatch.org}
  \item Laura Osborne \email{laura.osborne@globalfishingwatch.org}
  \item Andrea Sánchez-Tapia \email{andrea.sanchez-tapia@globalfishingwatch.org} (\href{https://orcid.org/0000-0002-3521-4338}{ORCID})
  \item Gavin McDonald \email{gmcdonald@bren.ucsb.edu} (\href{https://orcid.org/0000-0001-7624-5985}{ORCID})
}

Other contributors:
\itemize{
  \item Nate Miller \email{nate.miller@globalfishingwatch.org} [contributor]
}

}
\keyword{internal}
