% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_prep_data.r
\name{gear_correction}
\alias{gear_correction}
\title{Gear manual correction
Will need updating with more offenders and non offenders data}
\usage{
gear_correction(data)
}
\arguments{
\item{data}{data frame. Needs to have a gear (character) and ssvid (character) columns.}
}
\value{
data frame with corrected gears
}
\description{
Gear manual correction
Will need updating with more offenders and non offenders data
}
