% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classification.R
\name{ml_classification}
\alias{ml_classification}
\title{Computes binary classification via DEDPUL}
\usage{
ml_classification(
  data,
  steps = 1000,
  plotting = FALSE,
  filepath = NULL,
  threshold = seq(0, 0.99, by = 0.01),
  eps = 0.01,
  confidence_levels = TRUE,
  parallel_plan = "multicore",
  free_cores = 1
)
}
\arguments{
\item{data}{tibble with at least a common_seed column and a prediction_output
column The prediction_output column is a list. Each element contains a
tibble with predictions and covariates.}

\item{steps}{number of locations at which to compute D}

\item{plotting}{if TRUE, a D vs. alpha plot is generated}

\item{filepath}{if plotting is TRUE, a filepath of where to save the plot is
needed}

\item{threshold}{potential thresholds to test}

\item{eps}{accepted difference (tolerance) between alpha and the actual
proportion of positives for a given threshold}

\item{confidence_levels}{Boolean to compute confidence levels}

\item{parallel_plan}{type of parallelization to run (multicore, multisession,
or psock - this last one may need calling libraries inside)}

\item{free_cores}{number of free cores to leave out of parallelization}
}
\value{
tibble with classification and calibrated threshold used for them
}
\description{
For each vessel-year, it computes a binary
classification, 0 non offender and 1 offender. It is based on the DEDPUL
algorithm in the reference.
}
\references{
D. Ivanov, "DEDPUL: Difference-of-Estimated-Densities-based
Positive-Unlabeled Learning," 2020 19th IEEE International Conference on
Machine Learning and Applications (ICMLA), 2020, pp. 782-790, doi:
10.1109/ICMLA51294.2020.00128.
}
