% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_hyperpar.R
\name{ml_hyperpar}
\alias{ml_hyperpar}
\title{Get best hyperparameter combination for each common seed after ML training}
\usage{
ml_hyperpar(data)
}
\arguments{
\item{data}{data frame of train cross-validated datasets with
several bags, it must have columns:
.pred_1 : probability of being an offender;
bag: bag ID;
known_offender: 0 if not, 1 if yes;
.row: row ID
common_seed: common seed to generate bags}
}
\value{
data frame of best hyperparameter combinations per common seed
}
\description{
Get best hyperparameter combination for each common seed after ML training
}
