% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_perf_metrics.R
\name{ml_perf_metrics}
\alias{ml_perf_metrics}
\title{Computes recall for assessment sets and specificity for holdout non offenders}
\usage{
ml_perf_metrics(data)
}
\arguments{
\item{data}{tibble with at least a prediction_output column (\code{pred_class}),
a \code{holdout} column (whether if the observation was used in the model or
held out), a \code{known_offender} column (whether the vessel was identified as
an offender by reports), and a \code{known_non_offender} column (whether the
vessel was identified as non offender by inspections).}
}
\value{
tibble with recall and specificity per seed
}
\description{
Two performance metrics are computed:
recall, for assessment sets (in model versions that did not use them for
training), and specificity for holdout non offenders (in the same year of the
certification/inspection - if done at the end of the year)
}
