% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_pred_summary.R
\name{ml_pred_summary}
\alias{ml_pred_summary}
\title{Computes prediction (classification) summary between seeds}
\usage{
ml_pred_summary(data, num_common_seeds)
}
\arguments{
\item{data}{tibble with at least an indID (vessel-year ID) column, pred_class
column (predicted class in each seed),  thres column (threshold used for
classification with the corresponding common seed), and common_seed column.}

\item{num_common_seeds}{number of common seeds}
}
\value{
tibble with recall and specificity per seed
}
\description{
The mode of the class (0 or 1) is computed between seeds, and
the proportion of seeds matching that classification is given too.
}
