% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_prep_data.r
\name{ml_prep_data}
\alias{ml_prep_data}
\title{Preprocessing of offenders, possible offenders, non offenders and unlabeled
data}
\usage{
ml_prep_data(
  fl_data = fl_df$data,
  tidy_data = alltogether_tidy$data,
  gears_interest,
  vars_to_factor,
  vars_remove
)
}
\arguments{
\item{fl_data}{data frame of offenders. Needs to have columns such as: a gear
(character),
ssvid (character), engine_power_kw (double), tonnage_gt (double),
length_m (double), ais_type (character), event_ais_year (integer),
fl_event_id (integer), known_offender (double), known_non_offender (integer),
possible_offender (double)}

\item{tidy_data}{data frame of AIS. Needs to have the same columns as
fl_data.}

\item{gears_interest}{vector of character elements with names of the gears
of interest.}

\item{vars_to_factor}{vector of character elements with names of columns to
convert from character to factor. They have to be columns existing in fl_data
and tidy_data.}

\item{vars_remove:}{vector of character elements with names of columns to
remove from tidy_data.}
}
\value{
list with 2 elements:
holdout_set : data frame with AIS info from offenders before/after
the year of offense, potential offenders and known non offenders;
training set : data frame with AIS info from offenders during the year
of offense, and unlabeled cases
}
\description{
Preprocessing of offenders, possible offenders, non offenders and unlabeled
data
}
