% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_perf_metrics.R
\name{ml_recall}
\alias{ml_recall}
\title{Computes recall for assessment sets}
\usage{
ml_recall(data)
}
\arguments{
\item{data}{tibble with at least a prediction_output
column and a known_offender column (whether if the vessel was identified as
an offender by reports).}
}
\value{
recall value
}
\description{
Computes recall, for assessment sets (in model versions that
did not use them for training)
}
