% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_train_predict.R
\name{ml_train_predict}
\alias{ml_train_predict}
\title{Trains machine learning (RF) models and predicts}
\usage{
ml_train_predict(
  fl_rec,
  rf_spec,
  cv_splits_all,
  bag_runs,
  down_sample_ratio,
  parallel_plan = "multicore",
  free_cores = 1,
  prediction_df = NULL,
  save_dir = "models"
)
}
\arguments{
\item{fl_rec}{recipe}

\item{rf_spec}{model specifications}

\item{cv_splits_all}{tibble containing tibbles of cross-validation splits
(1 split per common seed)}

\item{bag_runs}{bags}

\item{down_sample_ratio}{down sampling ratio for the predicted class to add
to the recipe in each bag}

\item{parallel_plan}{type of parallelization to run (multicore, multisession,
or psock - this last one may need calling libraries inside)}

\item{free_cores}{number of free cores to leave out of parallelization}

\item{prediction_df}{hold-out data frame with possible offenders and non
offenders to predict on. If NULL (default), then only predict on the training set}
}
\value{
an object with predicted values and fitted models
}
\description{
For each bag seed, fit one random forest to each bag in each
fold, and predict over the assessment set and the holdout set.
}
