% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_training.R
\name{ml_training}
\alias{ml_training}
\title{Trains machine learning (RF) models:}
\usage{
ml_training(
  fl_rec,
  rf_spec,
  cv_splits_all,
  bag_runs,
  down_sample_ratio,
  num_grid = 5,
  parallel_plan = "multicore",
  free_cores = 1
)
}
\arguments{
\item{fl_rec}{recipe}

\item{rf_spec}{model specifications}

\item{cv_splits_all}{tibble containing tibbles of cross-validation splits.
1 cv split per common_seed}

\item{bag_runs}{data frame with the ID of each bag for each common seed and
the actual seed using to downsample in each bag}

\item{down_sample_ratio}{down sampling ratio for the smaller class
(offenders); to add to the recipe in each bag}

\item{num_grid}{number of grid random values for combinations of
hyperparameters per bag, or grid of values to test}

\item{parallel_plan}{type of parallelization to run (multicore, multisession
or psock - this last one may need calling libraries inside)}

\item{free_cores}{number of free cores to leave out of parallelization}
}
\value{
data frame with confidence scores of being an offender
}
\description{
For each bag seed, it generates a set of model hyperparameter
configurations,
}
