% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_median.R
\name{rolling_median}
\alias{rolling_median}
\title{Smoothing r using a rolling median}
\usage{
rolling_median(r, l_2 = 20)
}
\arguments{
\item{r}{sorted array of density ratios}

\item{l_2}{denominator to get rolling window of length(r)/l_2
(default to 20 based on the reference)}
}
\value{
sorted array of density ratios, smoothed
}
\description{
Smoothing r using a rolling median
}
\references{
D. Ivanov, "DEDPUL: Difference-of-Estimated-Densities-based
Positive-Unlabeled Learning," 2020 19th IEEE International Conference on
Machine Learning and Applications (ICMLA), 2020, pp. 782-790, doi:
10.1109/ICMLA51294.2020.00128.
}
