% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classification.R
\name{compute_D}
\alias{compute_D}
\title{Compute D from MAX_SLOPE algorithm}
\usage{
compute_D(r, steps = 1000)
}
\arguments{
\item{r}{sorted 1D array of density ratios}

\item{steps}{number of locations at which to compute D}
}
\value{
alpha and D vectors
}
\description{
Compute D from MAX_SLOPE algorithm
}
\details{
See Algorithm 2 in the reference.
\code{D = alpha - mean(p(Yu))} where \code{p(Yu)} is defined as
\code{p(Yu) = min(alpha * r(Yu), 1)}
}
\references{
D. Ivanov, "DEDPUL: Difference-of-Estimated-Densities-based
Positive-Unlabeled Learning," 2020 19th IEEE International Conference on
Machine Learning and Applications (ICMLA), 2020, pp. 782-790, doi:
10.1109/ICMLA51294.2020.00128.
}
