% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classification.R
\name{compute_r}
\alias{compute_r}
\title{Computes density ratios array r}
\usage{
compute_r(f_y)
}
\arguments{
\item{f_y}{list with f_yp, f_yu and y_u as elements (see details)}
}
\value{
sorted array of density ratios, monotonized and smoothed
}
\description{
Compute r using Algorithms 1 and 2 in reference.
}
\details{
In f_y, f_yp is the vector of inferred densities for positive
predictions, f_yu is the vector of inferred densities for unlabeled
predictions, and y_u is the vector with the predictions of unlabeled.
}
\references{
D. Ivanov, "DEDPUL: Difference-of-Estimated-Densities-based
Positive-Unlabeled Learning," 2020 19th IEEE International Conference on
Machine Learning and Applications (ICMLA), 2020, pp. 782-790, doi:
10.1109/ICMLA51294.2020.00128.
}
