% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_bag_downsample.R
\name{dev_bag_downsample}
\alias{dev_bag_downsample}
\title{Apply downsample step to data recipe across defined seeds/bags}
\usage{
dev_bag_downsample(bag_runs, fl_rec, down_sample_ratio)
}
\arguments{
\item{bag_runs}{Tibble defining bag numbers and seeds}

\item{fl_rec}{Model recipe}

\item{down_sample_ratio}{See under_ratio in ?themis::step_downsample. To reduce the weight of the unlabeled cases in the model, we randomly
downsampled them in the training set with a 1-1 ratio, i.e. the number of
positive and unlabeled cases used for training would be equal}
}
\value{
Tibble with data recipe and downsample ratio.
}
\description{
Apply downsample step to data recipe across defined seeds/bags
}
