% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_cv_setup.R
\name{dev_cv_setup}
\alias{dev_cv_setup}
\title{Setting up the data structure to train the RF model based on a number of folds, bags and common seeds.
If tune is set to TRUE, is provided  return..., otherwise return analysis/assessment splits and model workflows
across CV folds and bags.}
\usage{
dev_cv_setup(
  training_data,
  num_folds,
  num_bags,
  num_seeds,
  fl_rec,
  rf_spec,
  down_sample_ratio,
  tune = FALSE,
  tune_parameters = NULL,
  grid = NULL
)
}
\arguments{
\item{training_data}{Dataset over which to generate CV folds and tuning grid}

\item{num_folds}{Number of cross validation folds}

\item{num_bags}{Number of bags}

\item{num_seeds}{Number of common seeds}

\item{fl_rec}{Model recipe}

\item{rf_spec}{Random forest classifier specifications}

\item{down_sample_ratio}{See under_ratio in ?themis::step_downsample. To reduce the weight of the unlabeled cases in the model, we randomly
downsampled them in the training set with a 1-1 ratio, i.e. the number of
positive and unlabeled cases used for training would be equal}

\item{tune}{Boolean defining wether to perform hyperparameter tuning in rf_spec}

\item{tune_parameters}{String defining parameter (or parameters) over which to perform tuning. Default to NULL which perform tuning across all hyperparameters in rf_spec}

\item{grid}{Grid defining hyperparameters values in rf_spec over which to perform tuning. Must contain the same parameters specified in tune_parameters}
}
\value{
List object containing cv_folds (analysis/assessment), model workflow and seed/bag identifiers
}
\description{
Setting up the data structure to train the RF model based on a number of folds, bags and common seeds.
If tune is set to TRUE, is provided  return..., otherwise return analysis/assessment splits and model workflows
across CV folds and bags.
}
