% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_cv_setup.R
\name{dev_cv_setup}
\alias{dev_cv_setup}
\title{Defining cross-validation folds (analysis/assessment) and model workflows across bags.}
\usage{
dev_cv_setup(bag_runs, cv_splits_all, fl_rec, rf_spec, down_sample_ratio)
}
\arguments{
\item{bag_runs}{Tibble defining bag numbers and seeds}

\item{cv_splits_all}{Cross-validation data splits}

\item{fl_rec}{Recipes data recipe}

\item{rf_spec}{Random forest model specification}

\item{down_sample_ratio}{See under_ratio in ?themis::step_downsample}
}
\value{
List object containing cv_folds (analysis/assessment), model workflow and seed/bag identifiers
}
\description{
Defining cross-validation folds (analysis/assessment) and model workflows across bags.
}
