% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ml_train.R
\name{dev_ml_train}
\alias{dev_ml_train}
\title{Training forced labour random forest model.}
\usage{
dev_ml_train(cv_setup, free_cores, parallel_plan, save_dir)
}
\arguments{
\item{cv_setup}{List containing cv_folds (analysis/assessment), model workflow and seed/bag identifiers. Output from ?cv_setup}

\item{free_cores}{Number of available cores. Add more if you need to do many things at the same time}

\item{parallel_plan}{Parallelization strategy Options: multisession (if running RStudio), multicore (Linux, Mac and plain R) or psock (if multisession is not working well and you need to try something else)}

\item{save_dir}{Directory to save trained models}
}
\value{
List containing:
Trained random forest models
Tibble with predicted probabilities
}
\description{
Training forced labour random forest model.
}
