% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcedlabor-package.R
\docType{package}
\name{forcedlabor-package}
\alias{forcedlabor}
\alias{forcedlabor-package}
\title{forcedlabor: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GlobalFishingWatch/forcedlabor}
  \item Report bugs at \url{https://github.com/GlobalFishingWatch/forcedlabor/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
