% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcedlabor-package.R
\docType{package}
\name{forcedlabor-package}
\alias{forcedlabor}
\alias{forcedlabor-package}
\title{forcedlabor: forced labor machine learning algorithms}
\description{
forcedlabor is an R package containing the codes to identify forced labor on fishing vessels via a machine learning algorithm trained on a combination of known forced labor cases and AIS data predictors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GlobalFishingWatch/forcedlabor}
  \item Report bugs at \url{https://github.com/GlobalFishingWatch/forcedlabor/issues}
}

}
\author{
\strong{Maintainer}: Rocío Joo \email{rocio.joo@globalfishingwatch.org} (\href{https://orcid.org/0000-0003-0319-4210}{ORCID})

}
\keyword{internal}
