% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classification.R
\name{monotonize}
\alias{monotonize}
\title{Enforcing partial monotonicity on r}
\usage{
monotonize(r, y_u)
}
\arguments{
\item{r}{sorted array of density ratios}

\item{y_u}{vector with the predictions of unlabeled}
}
\value{
sorted array of density ratios, monotonized
}
\description{
Each element of r is forced to be monotonic where, for the
element, y_u > y_u.mean(). See Algorithm 2 in reference for more details.
}
\references{
D. Ivanov, "DEDPUL: Difference-of-Estimated-Densities-based
Positive-Unlabeled Learning," 2020 19th IEEE International Conference on
Machine Learning and Applications (ICMLA), 2020, pp. 782-790, doi:
10.1109/ICMLA51294.2020.00128.
}
